+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v4.4.0-dev (Dec 17, 2025)
+- [Bugfix]Ensure code_verifier can be optional in resetPassword call [#3567](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3567)
 
 ## v4.3.0 (Dec 17, 2025)